refactor(overlay): remove overlay helper
Migrate the deprecated popover to direct CDK overlay creation and keep its trigger parsing local. Retain arrow and scroll utilities as focused common helpers.

Closes #808

BREAKING CHANGE: remove the common overlay helper API

Use Angular CDK Overlay directly instead of getOverlay, getOverlayPositions, getPositionStrategy, makeOverlay, or makePositionStrategy. Keep trigger parsing local instead of using hasTrigger.
